use ExtendedTracer instead of Tracing and implement Tracer interface
